requirements.txt
Pillow==8.0.1
ttkthemes==3.1.1